Split views.py to multiple files
