+++ api/urls.py
@@ -1,11 +1,11 @@
 from django.urls import path
-from . import views
+from . import PostViews, AuthorViews, CategoryViews
 
 urlpatterns = [
-    path("posts/", views.post_list),
-    path("posts/<int:pk>/", views.post_detail),
-    path("categories/", views.category_list),
-    path("categories/<int:pk>/", views.category_detail),
-    path("authors/", views.author_list),
-    path("authors/<int:pk>/", views.author_detail),
+    path("posts/", PostViews.post_list),
+    path("posts/<int:pk>/", PostViews.post_detail),
+    path("categories/", CategoryViews.category_list),
+    path("categories/<int:pk>/", CategoryViews.category_detail),
+    path("authors/", AuthorViews.author_list),
+    path("authors/<int:pk>/", AuthorViews.author_detail),
 ]
